project/dataset_combining/requirements.txt
pillow
requests
jupyter
numpy
pandas
matplotlib
seaborn
graphviz
Cython
pycocotools
GitPython